requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database and ORM
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1

# ML and Embeddings
sentence-transformers==2.3.1
torch>=2.2.0

# LLM Client
openai==1.10.0

# Authentication & Security
PyJWT==2.8.0

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Monitoring & Metrics
prometheus-client==0.19.0

# Caching & Distributed State
redis[hiredis]==5.0.1

# Utilities
python-multipart==0.0.6
httpx==0.26.0

